LPS-67544 These plugins are not meant to be applied explicitly
