+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/LiferayBasePlugin.java
@@ -16,6 +16,8 @@ package com.liferay.gradle.plugins;
 
 import com.liferay.gradle.plugins.extensions.AppServer;
 import com.liferay.gradle.plugins.extensions.LiferayExtension;
+import com.liferay.gradle.plugins.internal.NodeDefaultsPlugin;
+import com.liferay.gradle.plugins.internal.SourceFormatterDefaultsPlugin;
 import com.liferay.gradle.plugins.internal.util.FileUtil;
 import com.liferay.gradle.plugins.internal.util.GradleUtil;
 import com.liferay.gradle.plugins.tasks.DirectDeployTask;
