+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/LiferayOSGiPlugin.java
@@ -21,6 +21,21 @@ import com.liferay.gradle.plugins.cache.CachePlugin;
 import com.liferay.gradle.plugins.css.builder.CSSBuilderPlugin;
 import com.liferay.gradle.plugins.extensions.LiferayExtension;
 import com.liferay.gradle.plugins.extensions.LiferayOSGiExtension;
+import com.liferay.gradle.plugins.internal.AlloyTaglibDefaultsPlugin;
+import com.liferay.gradle.plugins.internal.CSSBuilderDefaultsPlugin;
+import com.liferay.gradle.plugins.internal.EclipseDefaultsPlugin;
+import com.liferay.gradle.plugins.internal.IdeaDefaultsPlugin;
+import com.liferay.gradle.plugins.internal.JSModuleConfigGeneratorDefaultsPlugin;
+import com.liferay.gradle.plugins.internal.JSTranspilerDefaultsPlugin;
+import com.liferay.gradle.plugins.internal.JavadocFormatterDefaultsPlugin;
+import com.liferay.gradle.plugins.internal.JspCDefaultsPlugin;
+import com.liferay.gradle.plugins.internal.LangBuilderDefaultsPlugin;
+import com.liferay.gradle.plugins.internal.ServiceBuilderDefaultsPlugin;
+import com.liferay.gradle.plugins.internal.TLDFormatterDefaultsPlugin;
+import com.liferay.gradle.plugins.internal.TestIntegrationDefaultsPlugin;
+import com.liferay.gradle.plugins.internal.UpgradeTableBuilderDefaultsPlugin;
+import com.liferay.gradle.plugins.internal.WSDDBuilderDefaultsPlugin;
+import com.liferay.gradle.plugins.internal.XMLFormatterDefaultsPlugin;
 import com.liferay.gradle.plugins.internal.util.FileUtil;
 import com.liferay.gradle.plugins.internal.util.GradleUtil;
 import com.liferay.gradle.plugins.jasper.jspc.JspCPlugin;
