+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/internal/EclipseDefaultsPlugin.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.gradle.plugins;
+package com.liferay.gradle.plugins.internal;
 
+import com.liferay.gradle.plugins.BaseDefaultsPlugin;
 import com.liferay.gradle.plugins.internal.util.GradleUtil;
 
 import groovy.lang.Closure;
