+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/internal/TLDFormatterDefaultsPlugin.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.gradle.plugins;
+package com.liferay.gradle.plugins.internal;
 
+import com.liferay.gradle.plugins.BasePortalToolDefaultsPlugin;
 import com.liferay.gradle.plugins.tld.formatter.TLDFormatterPlugin;
 
 /**
