+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/internal/TestIntegrationDefaultsPlugin.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.gradle.plugins;
+package com.liferay.gradle.plugins.internal;
 
+import com.liferay.gradle.plugins.BaseDefaultsPlugin;
 import com.liferay.gradle.plugins.extensions.LiferayExtension;
 import com.liferay.gradle.plugins.extensions.TomcatAppServer;
 import com.liferay.gradle.plugins.internal.util.GradleUtil;
